web-app/requirements.txt
torch
torchvision
opencv-python
numpy
matplotlib
scikit-learn
kagglehub
pillow
tqdm
fastapi
uvicorn[standard]
python-multipart
